split examples to files and adjust readme in site
