+++ site/gatsby-node.js
@@ -1,9 +1,15 @@
 const {
-  resolveUrls,
-  addLanguageLinks,
-  addKontentTypeItemsLink,
+  resolveUrls
+} = require('./example-resolve-url-slugs');
+const {
+  addLanguageLinks
+} = require('./example-languages-link');
+const {
+  addKontentTypeItemsLink
+} = require('./example-type-items-link')
+const {
   linkUsedByContentItems,
-} = require("./examples")
+} = require("./example-used-by-content-item-link")
 
 exports.createSchemaCustomization = async api => {
   resolveUrls(api)
