+++ site/src/pages/index.js
@@ -1,6 +1,12 @@
 import React from "react"
 import { RichTextElement } from "@simply007org/gatsby-kontent-components"
+import { Link } from "gatsby"
 
 export default () => {
-  return <RichTextElement value="<div>Hello Kontent!</div>" />
+  return (
+    <>
+    <RichTextElement value="<div>Hello Kontent!</div>" />
+    <Link to="rich-text">Rich text element resolution showcase</Link>
+    </>
+  );
 }
